requirements.txt
MySQL-python==1.2.5
Django==1.11.5
djangorestframework==3.6.4
django-guardian==1.4.9